+++ internal/depinject/container.go
@@ -1,116 +1,80 @@
 package depinject
 
-// const (
-// 	buildErrorName   = "build"
-// 	resolveErrorName = "resolve"
-// )
+import (
+	"io"
+	"os"
 
-// // A Container is a dependency injection container.
-// type Container struct {
-// 	// The internal graph representation of the container.
-// 	graph *graph.DAG[*node.Node]
+	"github.com/skjdfhkskjds/depinject/internal/depinject3/types"
+	"github.com/skjdfhkskjds/depinject/internal/graph"
+)
 
-// 	// The node registry of the container.
-// 	registry *node.Registry
+type Container struct {
+	graph    *graph.DAG[*types.Node]
+	registry *types.Registry
 
-// 	// Whether the container requires sentinels.
-// 	hasIn  bool
-// 	hasOut bool
-// }
+	// The writer to dump the container's info to.
+	writer io.Writer
 
-// // NewContainer creates a new container.
-// func NewContainer() *Container {
-// 	return &Container{
-// 		graph:    graph.NewDAG[*node.Node](),
-// 		registry: node.NewRegistry(),
-// 		hasIn:    false,
-// 		hasOut:   false,
-// 	}
-// }
+	// Whether the container is ready to be invoked.
+	invokable bool
 
-// // build builds the container by iterating through every
-// // node, and creating edges in the internal graph representation
-// // based on the dependencies and outputs of each node.
-// func (c *Container) build() error {
-// 	// Before building, supply the sentinels.
-// 	if c.hasIn {
-// 		if err := c.supply(sentinels.In{}); err != nil {
-// 			return err
-// 		}
-// 	}
-// 	if c.hasOut {
-// 		if err := c.supply(sentinels.Out{}); err != nil {
-// 			return err
-// 		}
-// 	}
+	// Sorted nodes in topological order.
+	sortedNodes []*types.Node
 
-// 	for _, node := range c.registry.Nodes() {
-// 		for i, dep := range node.Dependencies() {
-// 			source, err := c.registry.Get(dep)
-// 			// If the last dependency is not found, and the node is variadic,
-// 			// we continue without error.
-// 			if err != nil && node.IsVariadic() && i == len(node.Dependencies())-1 {
-// 				continue
-// 			} else if err != nil {
-// 				return errors.New(err, buildErrorName, node.ID(), dep.Name())
-// 			}
+	// Options
+	// Instructs the container to enable the use of sentinel
+	// structs in constructor arguments and parses the struct's
+	// fields as constructor arguments.
+	useInSentinel bool
 
-// 			// Add the edge to the graph. If the edge violates
-// 			// the acyclicity constraint, return an error.
-// 			if err := c.graph.AddEdge(source, node); err != nil {
-// 				return errors.New(err, buildErrorName, node.ID(), dep.Name())
-// 			}
-// 		}
-// 	}
+	// Instructs the container to enable the use of sentinel
+	// structs in constructor outputs and parses the struct's
+	// fields as constructor outputs.
+	useOutSentinel bool
 
-// 	return nil
-// }
+	// Allows the container to match dependencies that are interfaces
+	// to types which are implementations of those interfaces.
+	inferInterfaces bool
 
-// // resolve resolves the container by iterating through every
-// // node in the container and executing them in a topological order.
-// func (c *Container) resolve() error {
-// 	order, err := c.graph.TopologicalSort()
-// 	if err != nil {
-// 		return err
-// 	}
+	// Allows the container to have multiple constructors with the same
+	// output type, and will process them as lists (slices or arrays).
+	inferLists bool
+}
 
-// 	for _, node := range order {
-// 		// Get the dependencies of the node.
-// 		depTypes := node.Dependencies()
-// 		deps := make([]any, 0, len(depTypes))
-// 		for i, dep := range depTypes {
-// 			source, err := c.registry.Get(dep)
-// 			// If the last dependency is not found, and the node is variadic,
-// 			// we continue without error.
-// 			if err != nil && node.IsVariadic() && i == len(node.Dependencies())-1 {
-// 				continue
-// 			} else if err != nil {
-// 				return errors.New(err, buildErrorName, node.ID(), dep.Name())
-// 			}
+// DefaultContainer returns a new container with the default options.
+func DefaultContainer() *Container {
+	return &Container{
+		writer:          os.Stdout,
+		invokable:       false,
+		sortedNodes:     nil,
+		useInSentinel:   false,
+		useOutSentinel:  false,
+		inferInterfaces: false,
+		inferLists:      false,
+	}
+}
 
-// 			value, err := source.ValueOf(dep)
-// 			if err != nil {
-// 				return errors.New(err, resolveErrorName, node.ID(), dep.Name())
-// 			}
+// NewContainer returns a new container with the given options.
+func NewContainer(opts ...Option) *Container {
+	c := DefaultContainer()
+	for _, opt := range opts {
+		opt(c)
+	}
 
-// 			// Append the underlying casted value to deps
-// 			deps = append(deps, value.Interface())
-// 		}
+	c.graph = graph.NewDAG[*types.Node]()
+	c.registry = types.NewRegistry(c.inferLists, c.inferInterfaces)
+	return c
+}
 
-// 		// Execute the node with the dependencies.
-// 		if err := node.Execute(deps...); err != nil {
-// 			return errors.New(err, resolveErrorName, node.ID())
-// 		}
-// 	}
+// Dump dumps the container's registry into the writer.
+func (c *Container) Dump() {
+	c.writer.Write([]byte(c.registry.Dump()))
+}
 
-// 	return nil
-// }
-
-// // addNode adds a node to the container.
-// func (c *Container) addNode(node *node.Node) error {
-// 	if err := c.graph.AddVertex(node); err != nil {
-// 		return err
-// 	}
-
-// 	return c.registry.Register(node)
-// }
+// Destroy destroys the container and frees its memory.
+func (c *Container) Destroy() {
+	c.graph = nil
+	c.registry = nil
+	c.sortedNodes = nil
+	c = nil
+}
