+++ internal/depinject/provide.go
@@ -1,43 +1,27 @@
 package depinject
 
-// import (
-// 	"github.com/skjdfhkskjds/depinject/internal/depinject/types/errors"
-// 	"github.com/skjdfhkskjds/depinject/internal/depinject/types/node"
-// )
+import "github.com/skjdfhkskjds/depinject/internal/depinject3/types"
 
-// const provideErrorName = "provide"
+const provideErrorName = "provide"
 
-// // Provide adds a set of constructors to the container.
-// // It returns an error if any of the constructors are invalid,
-// // or if adding them results in an invalid graph.
-// //
-// // Note: Constructors are added to the container in the order they are provided.
-// func (c *Container) Provide(constructors ...any) error {
-// 	var err error
-// 	for _, constructor := range constructors {
-// 		if err = c.provide(constructor); err != nil {
-// 			return err
-// 		}
-// 	}
-// 	return nil
-// }
+func (c *Container) Provide(constructors ...any) error {
+	for _, constructor := range constructors {
+		if err := c.provide(constructor); err != nil {
+			return err
+		}
+	}
+	return nil
+}
 
-// // provide adds a constructor to the container.
-// func (c *Container) provide(constructor any) error {
-// 	node, err := node.New(constructor)
-// 	if err != nil {
-// 		return err
-// 	}
+func (c *Container) provide(constructor any) error {
+	node, err := types.NewNode(constructor)
+	if err != nil {
+		return newContainerError(err, provideErrorName, node.ID())
+	}
 
-// 	// Handle sentinels for the node
-// 	if err = c.handleSentinelsForNode(node); err != nil {
-// 		return err
-// 	}
+	if err = c.register(node, provideErrorName); err != nil {
+		return newContainerError(err, provideErrorName, node.ID())
+	}
 
-// 	// Add the node to the graph
-// 	if err = c.addNode(node); err != nil {
-// 		return errors.New(err, provideErrorName, node.ID())
-// 	}
-
-// 	return nil
-// }
+	return nil
+}
